#17: moved List component directly in shop folder
