+++ src/components/shop/List.scss
@@ -1,4 +1,4 @@
-@import './../../variables';
+@import '../variables';
 
 .list {
   border-width: 3px;
